scripts/requirements.txt
feedparser
beautifulsoup4
python-dateutil
requests
lingua-language-detector
